+++ django/applications/catmaid/static/js/widgets/landmark-widget.js
@@ -48,6 +48,15 @@
     this.targeted3dViewerNames = new Set();
     // Whether to automatically interpolate between group transformations
     this.interpolateBetweenGroups = true;
+    // Whether to show landmark layers
+    this.showLandmarkLayers = true;
+    // Whether skeleton colors should be overridden
+    this.overrideColor = true;
+    // Override color and alpha, used if overrideColor is true.
+    this._overrideColor = '#3FFFD8';
+    this._overrideAlpha = 0.8;
+    // A node scaling factor to help distinguish nodes from regular ones.
+    this.nodeScaling = 1.5;
 
     // The current edit mode
     this.mode = 'landmarks';
@@ -62,6 +71,7 @@
   };
 
   LandmarkWidget.prototype = {};
+  LandmarkWidget.prototype.constructor = LandmarkWidget;
   $.extend(LandmarkWidget.prototype, new InstanceRegistry());
 
   LandmarkWidget.prototype.getName = function() {
@@ -188,6 +198,11 @@
     availableSources.forEach(function(sourceName) {
       this.removeDisplayFrom3dViewer(sourceName);
     }, this);
+
+    project.getStackViewers().forEach(function(sv) {
+      sv.removeLayer('landmarklayer');
+      sv.redraw();
+    });
   };
 
   /**
@@ -275,6 +290,8 @@
     // Update actual content
     let mode = LandmarkWidget.MODES[this.mode];
     mode.createContent(this.content, this);
+
+    this.updateLandmarkLayers();
   };
 
   LandmarkWidget.prototype.setMode = function(mode) {
@@ -504,6 +521,67 @@
     }
   };
 
+  // Get all relevant skeleton projection options
+  LandmarkWidget.prototype.getLandmarkLayerOptions = function() {
+    return {
+      "visible": this.showLandmarkLayers,
+      "scale": this.nodeScaling,
+      "overrideColor": this.overrideColor ? this._overrideColor : false,
+      "overrideAlpha": this.overrideColor ? this._overrideAlpha : false,
+    };
+  };
+
+  /**
+   * Makes sure all active landmark display transformations are shown as layers.
+   */
+  LandmarkWidget.prototype.updateLandmarkLayers = function() {
+    var options = this.getLandmarkLayerOptions();
+    // Create a skeleton projection layer for all stack viewers that
+    // don't have one already.
+    let transformations = new Set(this.displayTransformations);
+    project.getStackViewers().forEach(function(sv) {
+      var layer = sv.getLayer('landmarklayer');
+      if (options.visible) {
+        if (!layer) {
+          // Create new if not already present
+          layer = new CATMAID.LandmarkLayer(sv, options);
+          sv.addLayer('landmarklayer', layer);
+        }
+
+        // Update existing instance
+        let layerTransformations = new Set(layer.displayTransformations);
+        // Remove layer transformations that are not part of the widget
+        let nRemoved = 0;
+        for (let t of layerTransformations) {
+          if (!transformations.has(t)) {
+            layerTransformations.delete(t);
+            ++nRemoved;
+          }
+        }
+        // Add display transformations not yet part of layer
+        let nAdded = 0;
+        for (let t of transformations) {
+          if (!layerTransformations.has(t)) {
+            layerTransformations.add(t);
+            ++nAdded;
+          }
+        }
+
+        layer.displayTransformations = Array.from(layerTransformations);
+
+        if (nRemoved !== 0 || nAdded !== 0) {
+          layer.update();
+        }
+
+        // Update other options and display
+        layer.updateOptions(options, false, true);
+      } else if (layer) {
+        sv.removeLayer('landmarklayer');
+        sv.redraw();
+      }
+    });
+  };
+
   /**
    */
   LandmarkWidget.prototype.getMlsTransform = function(transformation, i) {
@@ -680,25 +758,33 @@
 
       transformation.nodeProvider = {
         get: function(skeletonId) {
-          if (transformation.skeletonCache && transformation.skeletonCache[skeletonId]) {
-            return Promise.resolve(transformation.skeletonCache[skeletonId]);
-          } else {
-            // Get skeleton data and transform it
-            return CATMAID.fetch(project.id + '/skeletons/' + skeletonId + '/compact-detail', 'GET', {
-                with_tags: false,
-                with_connectors: false,
-                with_history: false
-              })
-              .then(function(response) {
-                // Transform points and store in cache
-                response[0].forEach(transformTreenode);
-                if (!transformation.skeletonCache) {
-                  transformation.skeletonCache = {};
-                }
-                transformation.skeletonCache[skeletonId] = response;
-                return response;
-              });
+          if (!transformation.loading) {
+            if (transformation.skeletonCache && transformation.skeletonCache[skeletonId]) {
+              transformation.loading = Promise.resolve(transformation.skeletonCache[skeletonId]);
+            } else {
+              // Get skeleton data and transform it
+              transformation.loading = CATMAID.fetch(project.id + '/skeletons/' + skeletonId + '/compact-detail', 'GET', {
+                  with_tags: false,
+                  with_connectors: false,
+                  with_history: false
+                })
+                .then(function(response) {
+                  // If the source group ID is the same as the target group ID,
+                  // don't transform at all.
+                  if (transformation.fromGroupId !== transformation.toGroupId) {
+                    // Transform points and store in cache
+                    response[0].forEach(transformTreenode);
+                  }
+                  if (!transformation.skeletonCache) {
+                    transformation.skeletonCache = {};
+                  }
+                  transformation.skeletonCache[skeletonId] = response;
+                  return response;
+                });
+            }
           }
+
+          return transformation.loading;
         }
       };
 
@@ -1649,6 +1735,18 @@
         let target3dViewerSelect = document.createElement('span');
         target3dViewerSelect.setAttribute('data-role', 'display-target');
         target.updateTargetSelect(target3dViewerSelect);
+
+        let colorButton = document.createElement('span');
+        CATMAID.DOM.appendColorButton(colorButton, 'c',
+            'Override color for landmark display',
+            undefined, function(colorRGB, alpha, colorChanged, alphaChanged, colorHex) {
+              target._overrideColor = '#' + colorHex;
+              target._overrideAlpha = alpha;
+              target.updateLandmarkLayers();
+            }, {
+              initialColor: target._overrideColor,
+              initialAlpha: target._overrideAlpha
+            });
         return [
           {
             type: 'button',
@@ -1676,6 +1774,41 @@
               target.interpolateBetweenGroups = this.checked;
               target.updateDisplay();
             }
+          },
+          {
+            type: 'checkbox',
+            value: target.showLandmarkLayers,
+            label: 'Show landmark layers',
+            onclick: function() {
+              target.showLandmarkLayers = this.checked;
+              target.updateLandmarkLayers();
+            }
+          },
+          {
+            type: 'numeric',
+            value: target.nodeScaling,
+            label: 'Node scaling',
+            length: 3,
+            onchange: function() {
+              let val = parseFloat(this.value);
+              if (!Number.isNaN(val)) {
+                target.nodeScaling = val;
+                target.updateLandmarkLayers();
+              }
+            }
+          },
+          {
+            type: 'checkbox',
+            value: target.overrideColor,
+            label: 'Override color',
+            onclick: function() {
+              target.overrideColor = this.checked;
+              target.updateLandmarkLayers();
+            }
+          },
+          {
+            type: 'child',
+            element: colorButton
           }
         ];
       },
@@ -1851,7 +1984,33 @@
     name: "Landmarks",
     description: "Show and manage landmark locations",
     key: "landmarks",
-    creator: LandmarkWidget
+    creator: LandmarkWidget,
+    state: {
+      getState: function(widget) {
+        return {
+          importAllowNonEmptyGroups: widget.importAllowNonEmptyGroups,
+          importCreateNonExistingGroups: widget.importCreateNonExistingGroups,
+          importReuseExistingLandmarks: widget.importReuseExistingLandmarks,
+          interpolateBetweenGroups: widget.interpolateBetweenGroups,
+          showLandmarkLayers: widget.showLandmarkLayers,
+          overrideColor: widget.overrideColor,
+          overrideColorHex: widget._overrideColor,
+          overrideColorAlpha: widget._overrideAlpha,
+          nodeScaling: widget.nodeScaling
+        };
+      },
+      setState: function(widget, state) {
+        CATMAID.tools.copyIfDefined(state, widget, 'importReuseExistingLandmarks');
+        CATMAID.tools.copyIfDefined(state, widget, 'importCreateNonExistingGroups');
+        CATMAID.tools.copyIfDefined(state, widget, 'importReuseExistingLandmarks');
+        CATMAID.tools.copyIfDefined(state, widget, 'interpolateBetweenGroups');
+        CATMAID.tools.copyIfDefined(state, widget, 'showLandmarkLayers');
+        CATMAID.tools.copyIfDefined(state, widget, 'overrideColor');
+        CATMAID.tools.copyIfDefined(state, widget, 'overrideColorHex');
+        CATMAID.tools.copyIfDefined(state, widget, 'overrideColorAlpha');
+        CATMAID.tools.copyIfDefined(state, widget, 'nodeScaling');
+      }
+    }
   });
 
 })(CATMAID);